reorganized the math functions with numpy categories, more intuitive rounding
