+++ toolkits/math/evals/eval_math_tools.py
@@ -9,27 +9,39 @@ from arcade.sdk.eval import (
 
 import arcade_math
 from arcade_math.tools.arithmetic import (
-    abs_val,
     add,
-    avg,
-    ceil,
-    deg_to_rad,
     divide,
-    factorial,
-    floor,
-    gcd,
-    lcm,
-    log,
-    median,
     mod,
     multiply,
-    power,
-    rad_to_deg,
-    round_num,
-    sqrt,
     subtract,
     sum_list,
     sum_range,
+)
+from arcade_math.tools.exponents import (
+    log,
+    power,
+)
+from arcade_math.tools.miscellaneous import (
+    abs_val,
+    factorial,
+    sqrt,
+)
+from arcade_math.tools.rational import (
+    gcd,
+    lcm,
+)
+from arcade_math.tools.rounding import (
+    ceil,
+    floor,
+    round_num,
+)
+from arcade_math.tools.statistics import (
+    avg,
+    median,
+)
+from arcade_math.tools.trigonometry import (
+    deg_to_rad,
+    rad_to_deg,
 )
 
 # Evaluation rubric
